feat(home): fully commit to use emacs as a daemon
Note that the first two connections to the emacs daemon will produce
broken frames (with teeny tiny fonts), but overall is an improvement.

(We'll try to fix that issue, now that we are going to face it every
restart of the daemon).

Thanks to @CesarGallego for leading the way on this one!