feat(home): hyprland: launch fuzzel with SUPER D
Following Gnome style of launching the launcher with a single press to
the Hyper button was not working so well for me, even though I thought I
would love it.  But I find that more often than not, I press the key
without meaning to, and I end up with a blocking fuzzel that I have to
take care of.  But reverting to SUPER D, I now only launch fuzzen when I
really mean to.