chore(flake): unpin Hyprland
After adding `programs.hyprland.enable = true;` to my
`configuration.nix`, the issue is fixed for me.

Without it, Hyprland NixOS module was not being properly used, and a
crucial portal setup was not being made.  This became critical with the
commit following the pinned version, which depended on the env var
`XDG_DESKTOP_PORTAL_DIR` being defined, the setting of which didn't happen without this
setup.